+++ src/prompt/model.py
@@ -69,3 +69,38 @@ class OpenCLIP(nn.Module):
         elif dtype == 'text':
             feat = self.openclip.encode_text(data)
         return feat
+
+
+class PromptCLIP(nn.Module):
+    def __init__(self, model_args, tgt_device='cpu'):
+        super(PromptCLIP, self).__init__()
+        self.model_args = model_args
+        self.openclip, self.pre_process_train, self.pre_process_val = open_clip.create_model_and_transforms(
+            model_name='ViT-L-14', pretrained='laion2b_s32b_b82k', device=tgt_device,
+        )
+        self.tokenizer = open_clip.get_tokenizer('ViT-L-14')
+
+        self.embedding_layer = nn.Sequential(*list(self.openclip.visual.children())[:3])
+        self.transformer_layer = nn.Sequential(*list(self.openclip.visual.children())[4:])
+        self.prompt = nn.Parameter(torch.randn(
+            self.model_args.n_prompts, self.model_args.prompt_dim))
+    
+        self.triplet_loss = nn.TripletMarginWithDistanceLoss(
+            distance_function=lambda x, y: 1.0-F.cosine_similarity(x, y), 
+            margin=1)
+        
+        
+    
+    def forward(self, data, dtype='image'):
+        if dtype == 'image': 
+            embed = self.embedding_layer(data)
+            embed = embed.flatten(2).transpose(-1, -2)
+            embed = torch.cat((
+                self.prompt.expand(data.shape[0],-1,-1),
+                embed,
+            ), dim=1)
+            feat = self.transformer_layer(embed)
+
+        elif dtype == 'text':
+            feat = self.openclip.encode_text(data)
+        return feat